Move types and auxiliary routines into separate headers
